Show delete option only for Playlist.Smart & FavouritePlayList

+++ mobile/src/main/java/com/kingbull/musicplayer/ui/main/categories/playlists/members/MemberQuickAction.java
@@ -12,33 +12,32 @@ import com.kingbull.musicplayer.ui.base.musiclist.quickaction.QuickAction;
  * @author Kailash Dabhi
  * @date 12/8/2016.
  */
-
 public final class MemberQuickAction {
-
   private final int ID_PLAY = 1;
   private final int ID_MOVE_TO = 2;
   private final int ID_EDIT_TAGS = 3;
   private final int ID_RINGTONE = 4;
   private final int ID_DELETE = 5;
   private final int ID_SEND = 6;
-
   private Activity activity;
   private QuickAction quickAction;
   private MemberQuickActionListener memberQuickActionListener;
+  private boolean hasDeleteOption;
 
-  public MemberQuickAction(final Activity activity) {
+  public MemberQuickAction(final Activity activity, boolean hasDeleteOption) {
     this.activity = activity;
+    this.hasDeleteOption = hasDeleteOption;
     final ActionItem playItem = new ActionItem(ID_PLAY, "Play",
         ContextCompat.getDrawable(activity, R.drawable.composer_button_play));
     final ActionItem addToPlaylistItem = new ActionItem(ID_MOVE_TO, "Move To",
         ContextCompat.getDrawable(activity, R.drawable.composer_button_multiselect));
-    ActionItem editTagsItem = new ActionItem(ID_EDIT_TAGS, "Edit Tags",
+    final ActionItem editTagsItem = new ActionItem(ID_EDIT_TAGS, "Edit Tags",
         ContextCompat.getDrawable(activity, R.drawable.composer_button_sort));
     final ActionItem setAsRingtoneItem = new ActionItem(ID_RINGTONE, "Set As Ringtone",
         ContextCompat.getDrawable(activity, R.drawable.composer_button_shuffle));
     final ActionItem deleteItem = new ActionItem(ID_DELETE, "Delete from playlist",
         ContextCompat.getDrawable(activity, R.drawable.composer_button_delete));
-    ActionItem sendItem = new ActionItem(ID_SEND, "Send",
+    final ActionItem sendItem = new ActionItem(ID_SEND, "Send",
         ContextCompat.getDrawable(activity, R.drawable.composer_button_queue));
     //use setSticky(true) to disable QuickAction dialog being dismissed after an item is clicked
     quickAction = new QuickAction(activity);
@@ -70,7 +69,9 @@ public final class MemberQuickAction {
     quickAction.addActionItem(addToPlaylistItem);
     quickAction.addActionItem(editTagsItem);
     quickAction.addActionItem(setAsRingtoneItem);
-    quickAction.addActionItem(deleteItem);
+    if (hasDeleteOption) {
+      quickAction.addActionItem(deleteItem);
+    }
     quickAction.addActionItem(sendItem);
   }
